Criação dos endpoints e toda estrutura principal

+++ src/main/java/aplicacao/proj/rest/dto/simulacao/SimulacaoRequestDto.java
@@ -1,8 +1,10 @@
 package aplicacao.proj.rest.dto.simulacao;
 
+import java.math.BigDecimal;
+
 public record SimulacaoRequestDto(
-        Long clienteId,
-        Double valor,
+        Integer clienteId,
+        BigDecimal valor,
         Integer prazoMeses,
         String tipoProduto
 ) {}
